+++ geotrek/zoning/tests/test_parsers.py
@@ -10,6 +10,7 @@ from django.test import TestCase
 from geotrek.zoning.models import City, District, RestrictedArea, RestrictedAreaType
 from geotrek.zoning.parsers import (
     CityParser,
+    OpenStreetMapCityParser,
     OpenStreetMapDistrictParser,
     OpenStreetMapRestrictedAreaParser,
 )
@@ -279,4 +280,118 @@ class OpenStreetMapRestrictedAreaParserTests(TestCase):
         self.assertEqual(
             self.restricted_areas[0].name,
             "Parc Naturel Régional des Pyrénées Ariégeoises",
+        )
+
+
+class TestCityOpenStreetMapParser(OpenStreetMapCityParser):
+    provider = "OpenStreetMap"
+    tags = [[{"boundary": "administrative"}, {"admin_level": "8"}]]
+    code_tag = "ref:INSEE"
+
+
+class OpenStreetMapCityParserTests(TestCase):
+    @mock.patch("geotrek.common.parsers.requests.get")
+    def import_cities(self, nominatim_file, status_code, mocked):
+        def mocked_json_overpass():
+            filename = os.path.join(os.path.dirname(__file__), "data", "city_OSM.json")
+            with open(filename) as f:
+                return json.load(f)
+
+        def mocked_json_nominatim():
+            filename = os.path.join(os.path.dirname(__file__), "data", nominatim_file)
+            with open(filename) as f:
+                return json.load(f)
+
+        response1 = mock.Mock()
+        response1.json = mocked_json_overpass
+        response1.status_code = 200
+
+        response2 = mock.Mock()
+        response2.json = mocked_json_nominatim
+        response2.status_code = status_code
+        response2.url = (
+            "https://nominatim.openstreetmap.org/ui/details.html?osmtype=R&osmid=68921"
+        )
+
+        mocked.side_effect = [response1, response2]
+
+        output = StringIO()
+        call_command(
+            "import",
+            "geotrek.zoning.tests.test_parsers.TestCityOpenStreetMapParser",
+            verbosity=2,
+            stdout=output,
+        )
+
+        self.output = output.getvalue()
+        self.cities = City.objects.order_by("pk").all()
+
+    def test_multipolygon_OSM(self):
+        nominatim_file = os.path.join(
+            os.path.dirname(__file__), "data", "OSM_multipolygon.json"
+        )
+        self.import_cities(nominatim_file, 200)
+
+        self.assertEqual(self.cities.count(), 1)
+        self.assertEqual(type(self.cities[0].geom), MultiPolygon)
+        self.assertEqual(len(self.cities[0].geom), 2)
+
+        # test the first point of each polygon
+        self.assertAlmostEqual(
+            self.cities[0].geom.coords[0][0][0][0], 869953.433, places=2
+        )
+        self.assertAlmostEqual(
+            self.cities[0].geom.coords[0][0][0][1], 6365109.915, places=2
+        )
+
+        self.assertAlmostEqual(
+            self.cities[0].geom.coords[1][0][0][0], 872596.909, places=2
+        )
+        self.assertAlmostEqual(
+            self.cities[0].geom.coords[1][0][0][1], 6367024.025, places=2
+        )
+
+    def test_polygon_OSM(self):
+        nominatim_file = os.path.join(
+            os.path.dirname(__file__), "data", "OSM_polygon.json"
+        )
+        self.import_cities(nominatim_file, 200)
+
+        self.assertEqual(self.cities.count(), 1)
+        self.assertEqual(type(self.cities[0].geom), MultiPolygon)
+        self.assertEqual(len(self.cities[0].geom), 1)
+
+        # test the first point of each polygon
+        self.assertAlmostEqual(
+            self.cities[0].geom.coords[0][0][0][0], 872596.909, places=2
+        )
+        self.assertAlmostEqual(
+            self.cities[0].geom.coords[0][0][0][1], 6367024.025, places=2
+        )
+
+    def test_geom_does_not_exist(self):
+        nominatim_file = os.path.join(
+            os.path.dirname(__file__), "data", "OSM_polygon.json"
+        )
+        self.import_cities(nominatim_file, 404)
+
+        self.assertEqual(self.cities.count(), 0)
+
+        self.assertIn(
+            "Failed to download https://nominatim.openstreetmap.org/ui/details.html?osmtype=R&osmid=68921. HTTP status code 404",
+            self.output,
+        )
+
+    def test_create_OSM(self):
+        nominatim_file = os.path.join(
+            os.path.dirname(__file__), "data", "OSM_polygon.json"
+        )
+        self.import_cities(nominatim_file, 200)
+
+        self.assertEqual(self.cities.count(), 1)
+
+        self.assertEqual(self.cities[0].name, "Paizay-le-Sec")
+        self.assertEqual(
+            self.cities[0].code,
+            "86187",
         )